Refactor email templates and implement naming conventions for improved consistency
