+++ packages/emails/src/templates/index.ts
@@ -1,0 +1,1 @@
+export * from '@/templates/failed-backup.js';
