+++ packages/emails/src/types.ts
@@ -1,0 +1,6 @@
+/* * */
+
+export interface SendEmailProps<T> {
+	data: T
+	to: string | string[]
+}
