+++ packages/eslint/src/rules/naming-conventions.ts
@@ -1,0 +1,69 @@
+/* * */
+
+import { type Config } from 'eslint/config';
+
+/**
+ * Enforces consistent naming conventions across all file types.
+ */
+export const namingConventionsConfig: Config[] = [
+
+	{
+		// files: ['**/*.*'], // Leave disable to apply to all file types.
+		rules: {
+			'@typescript-eslint/naming-convention': [
+				'error',
+				// Variables and functions: camelCase
+				{
+					format: ['camelCase'],
+					leadingUnderscore: 'allow',
+					selector: 'variableLike',
+				},
+				{
+					format: ['camelCase'],
+					selector: 'function',
+				},
+				// Constants: SCREAMING_SNAKE_CASE
+				{
+					format: ['UPPER_CASE', 'camelCase', 'PascalCase'], // Allow both for flexibility
+					modifiers: ['const', 'global'],
+					selector: 'variable',
+				},
+				// Types and interfaces: PascalCase
+				{
+					format: ['PascalCase'],
+					selector: 'typeLike',
+				},
+				// Class members: camelCase
+				{
+					format: ['camelCase'],
+					selector: 'classMethod',
+				},
+				{
+					format: ['camelCase'],
+					leadingUnderscore: 'allow',
+					selector: 'classProperty',
+				},
+				// Enum members: PascalCase or UPPER_CASE
+				{
+					format: ['PascalCase', 'UPPER_CASE'],
+					selector: 'enumMember',
+				},
+			],
+		},
+	},
+
+	{
+		files: ['**/*.tsx'],
+		rules: {
+			'@typescript-eslint/naming-convention': [
+				'error',
+				{
+					format: ['PascalCase'],
+					leadingUnderscore: 'allow',
+					selector: 'function',
+				},
+			],
+		},
+	},
+
+];
