+++ packages/eslint/src/sets/next.ts
@@ -88,41 +88,8 @@ export default [
 			'react-hooks/exhaustive-deps': 'warn',
 			'react-hooks/rules-of-hooks': 'error',
 			'react/jsx-key': 'error',
-
 			'react/no-children-prop': 'error',
 			'react/no-unescaped-entities': 'error',
-
-			// Frontend-specific naming conventions
-			'@typescript-eslint/naming-convention': [
-				'error',
-				// React Components: PascalCase
-				{
-					filter: {
-						match: true,
-						regex: '^[A-Z]', // Functions starting with capital (React components)
-					},
-					format: ['PascalCase'],
-					selector: 'function',
-				},
-				// React hooks: camelCase starting with 'use'
-				{
-					filter: {
-						match: true,
-						regex: '^use[A-Z]',
-					},
-					format: ['camelCase'],
-					selector: 'function',
-				},
-				// Props interfaces: PascalCase ending with Props
-				{
-					filter: {
-						match: true,
-						regex: 'Props$',
-					},
-					format: ['PascalCase'],
-					selector: 'interface',
-				},
-			],
 		},
 		settings: {
 			react: {
